# ff-split/pkg/api/openapi.yaml
openapi: 3.0.3
info:
  title: FinFlow Split Service API
  description: |
    API для сервиса управления совместными расходами в FinFlow.
    
    Сервис предоставляет функциональность для:
    - Управления мероприятиями и событиями
    - Создания и отслеживания транзакций
    - Разделения расходов между пользователями
    - Управления задачами и активностями
    - Работы с категориями и иконками
    
    ## Аутентификация
    
    Защищенные endpoints требуют аутентификации через Bearer токен:
    ```
    Authorization: Bearer <access_token>
    ```
  version: 1.0.0
  contact:
    name: FinFlow Team
  license:
    name: MIT
servers:
  - url: http://localhost:8081
    description: Development server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/event/{id_event}/activity:
    get:
      tags:
        - activities
      summary: Получить активности мероприятия
      description: Возвращает список активностей мероприятия
      operationId: getActivitiesByEventID
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список активностей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityListResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags:
        - activities
      summary: Создать активность
      description: Создает новую активность в мероприятии
      operationId: createActivity
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityRequest'
      responses:
        '201':
          description: Активность создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/event/{id_event}/activity/{id_activity}:
    get:
      tags:
        - activities
      summary: Получить активность
      description: Возвращает информацию об активности
      operationId: getActivityByID
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id_activity
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация об активности
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '404':
          description: Активность не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      tags:
        - activities
      summary: Обновить активность
      description: Обновляет информацию об активности
      operationId: updateActivity
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id_activity
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityRequest'
      responses:
        '200':
          description: Активность обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Активность не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      tags:
        - activities
      summary: Удалить активность
      description: Удаляет активность
      operationId: deleteActivity
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id_activity
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Активность удалена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Активность не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/category:
    get:
      tags:
        - categories
      summary: Получить список категорий
      description: Возвращает список всех категорий указанного типа
      operationId: getCategories
      parameters:
        - name: category_type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CategoryType'
          description: Тип категории
      responses:
        '200':
          description: Список категорий
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/category/{id}:
    get:
      tags:
        - categories
      summary: Получить категорию
      description: Возвращает информацию о категории по ID и типу
      operationId: getCategoryByID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID категории
        - name: category_type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CategoryType'
          description: Тип категории
      responses:
        '200':
          description: Информация о категории
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          description: Категория не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/manage/category:
    post:
      tags:
        - categories
      summary: Создать категорию
      description: Создает новую категорию указанного типа
      operationId: createCategory
      parameters:
        - name: category_type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CategoryType'
          description: Тип категории
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Категория создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/manage/category/{id}:
    put:
      tags:
        - categories
      summary: Обновить категорию
      description: Обновляет информацию о категории указанного типа
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID категории
        - name: category_type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CategoryType'
          description: Тип категории
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: Категория обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Категория не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      tags:
        - categories
      summary: Удалить категорию
      description: Удаляет категорию указанного типа
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID категории
        - name: category_type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CategoryType'
          description: Тип категории
      responses:
        '200':
          description: Категория удалена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Категория не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/manage/icons/{id}:
    get:
      tags:
        - icons
      summary: Получить иконку
      description: Возвращает информацию об иконке по ID
      operationId: getIconByID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация об иконке
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IconDTO'
        '404':
          description: Иконка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      tags:
        - icons
      summary: Обновить иконку
      description: Обновляет информацию об иконке
      operationId: updateIcon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IconRequest'
      responses:
        '200':
          description: Иконка обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IconDTO'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Иконка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      tags:
        - icons
      summary: Удалить иконку
      description: Удаляет иконку
      operationId: deleteIcon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Иконка удалена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Иконка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Авторизация через Bearer токен
  schemas:
    ErrorResponse:
      type: object
      required:
        - id
        - error
      properties:
        id:
          type: string
          description: ID запроса
        error:
          $ref: '#/components/schemas/ErrorResponseDetail'
    ErrorResponseDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Код ошибки
          example: "validation"
        message:
          type: string
          description: Сообщение об ошибке
          example: "invalid input"
        data:
          $ref: '#/components/schemas/ErrorResponseAdditionalData'
    ErrorResponseAdditionalData:
      type: object
      properties:
        slug:
          type: string
          description: Дополнительная информация
          example: "slug"
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Статус успеха операции
...
    ActivityRequest:
      type: object
      required:
        - description
      properties:
        description:
          type: string
          description: Описание активности
        icon_id:
          type: integer
          description: ID иконки
        user_id:
          type: integer
          format: int64
          description: Внутренний ID пользователя
    ActivityResponse:
      type: object
      properties:
        activity_id:
          type: integer
          description: ID активности
        description:
          type: string
          description: Описание активности
        icon_id:
          type: integer
          description: ID иконки
        datetime:
          type: string
          format: date-time
          description: Дата и время активности
    ActivityListResponse:
      type: object
      properties:
        activities:
          type: array
          items:
            $ref: '#/components/schemas/ActivityResponse'
...
    CategoryRequest:
      type: object
      required:
        - name
        - icon_id
      properties:
        name:
          type: string
          description: Название категории
        icon_id:
          type: integer
          description: ID иконки
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          description: ID категории
        name:
          type: string
          description: Название категории
        icon:
          $ref: '#/components/schemas/IconDTO'
    CategoryListResponse:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryResponse'
    IconRequest:
      type: object
      required:
        - name
        - file_uuid
      properties:
        name:
          type: string
          description: Название иконки
        file_uuid:
          type: string
          description: UUID файла иконки
    IconDTO:
      type: object
      properties:
        id:
          type: integer
          description: ID иконки
        name:
          type: string
          description: Название иконки
        file_uuid:
          type: string
          description: UUID файла иконки
        external_uuid:
          type: string
          description: Внешний UUID
    CategoryType:
      type: string
      enum:
        - event
        - transaction
      description: Тип категории